+++ client/src/components/rows/TrackRow.module.css
@@ -3,10 +3,15 @@
   align-items: center;
   height: 5rem;
   border-bottom: 1px solid #333;
+  padding: 0 4rem;
 }
 
 .trackRow:last-child {
   border-bottom: none;
+}
+
+.trackRow:hover {
+  background-color: rgba(0, 157, 224, 0.1);
 }
 
 .trackPlay {
@@ -18,6 +23,7 @@
   border: 0;
   margin-right: 1rem;
   padding: 0;
+  cursor: pointer;
 }
 
 .trackTitle {
@@ -26,4 +32,16 @@
 
 .trackArtist {
   color: #767676;
+}
+
+.addButton {
+  width: 2rem;
+  height: 2rem;
+  border-radius: 50%;
+  background: #333;
+  border: none;
+  box-shadow: 0;
+  padding: 0;
+  cursor: pointer;
+  margin-left: auto;
 }
