+++ client/src/components/tabs/PlaylistsTab.module.css
@@ -1,0 +1,8 @@
+.createButton {
+    border: 1px solid #333;
+    font-weight: 600;
+    color: #fff;
+    background-color: #333333;
+    padding: 1rem 2rem;
+    cursor: pointer;
+}
